+++ Library/src/test/java/dev/testify/ScreenshotRuleTest.kt
@@ -178,7 +178,7 @@ class ScreenshotRuleTest {
         every { pauseForInspection() } just runs
         every { HighContrastDiff.create(any(), any()) } returns mockHighContrastDiff
         every { mockEspressoHelper.syncUiThread() } just runs
-        every { closeSoftKeyboard() } just runs
+        every { closeSoftKeyboard(any()) } just runs
 
         val slot = slot<Runnable>()
         every { mockActivity.runOnUiThread(capture(slot)) } answers {